+++ src/Config/Hint/DefinitionHintInterface.php
@@ -3,14 +3,16 @@ declare(strict_types=1);
 
 namespace WoohooLabs\Zen\Config\Hint;
 
+use WoohooLabs\Zen\Config\EntryPoint\EntryPointInterface;
 use WoohooLabs\Zen\Container\Definition\DefinitionInterface;
 
 interface DefinitionHintInterface
 {
     /**
+     * @param EntryPointInterface[] $entryPoints
      * @param DefinitionHintInterface[] $definitionHints
      * @return DefinitionInterface[]
      * @internal
      */
-    public function toDefinitions(array $definitionHints, string $id, bool $isEntryPoint, bool $isAutoloaded, bool $isFileBased): array;
+    public function toDefinitions(array $entryPoints, array $definitionHints, string $id, bool $isAutoloaded, bool $isFileBased): array;
 }